% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STHD-RR-Main.R
\name{SthdRR}
\alias{SthdRR}
\title{The main entry point}
\usage{
SthdRR(FilePath)
}
\arguments{
\item{FilePath}{"./" is default, but if running from different directory, this is the path.}
}
\description{
Using the given FilePath, run the Steelhead Run Reconstruction
}
